more changes to rubber

+++ templates/xtrabackup/config/rubber/role/db/rubber-xtrarestore.sh
@@ -5,15 +5,48 @@
 %>#!/bin/bash
 # Variables
 LOGFILE="/tmp/rubber-xtrarestore-log"
-# Take the file from STDIN, write it into a real file, extract it, service mysql stop,
-# then mkdir <%=rubber_env.mount_directory%>/mysql/data & <%=rubber_env.mount_directory%>/mysql/log (move old ones out of the way)
-# then innobackupex --copy-back . in the extracted folder, then service mysql start
-# Create a temporary folder
+
+# the buckets on S3 where database backups are stored
+BACKUP_BUCKET="<%=rubber_env.cloud_providers.aws.backup_bucket%>"
+
+# Lets get our command line parameter
+while getopts ":u:p:t:db:" opt; do
+  case $opt in
+    t)
+      BACKUPFILE="$OPTARG"
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      exit 1
+      ;;
+    :)
+      echo "Option -$OPTARG requires an argument." >&2
+      exit 1
+      ;;
+  esac
+done
+
+if [[ -z "$BACKUPFILE" ]]; then
+  echo "Required parameters missing. Please supply -t (backup file to use to restore database)"
+  exit 1
+fi
+
+# Take the filename from STDIN and download it from amazon s3
+
 rm -rf <%=rubber_env.mount_directory%>/db_restore
 mkdir -p <%=rubber_env.mount_directory%>/db_restore
-# Write STDIN into file
-cat > <%=rubber_env.mount_directory%>/db_restore/current.tar.gz
 cd <%=rubber_env.mount_directory%>/db_restore
+
+# Download file from amazon s3
+#cat > /mnt/archon/db_restore/current.tar.gz
+s3cmd get s3://logix.cz-test/addrbook.xml addressbook-2.xml
+s3cmd get --config="<%=rubber_env.mount_directory%>/archon/db-backup-tools/rubber-s3cmd.s3cfg" "s3://$BACKUP_BUCKET/db/$BACKUPFILE" "current.tar.gz"
+
+# extract the downloaded file, service mysql stop,
+# then mkdir /mnt/archon/mysql/data & /mnt/archon/mysql/log (move old ones out of the way)
+# then innobackupex --copy-back . in the extracted folder, then service mysql start
+# Create a temporary folder
+
 tar xzvf current.tar.gz
 echo 'Stopping MySQL'
 if [ -z "`service mysql stop | grep 'done'`" ] ; then
